requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
boto==2.49.0
    # via pysparkling
certifi==2020.12.5
    # via requests
chardet==4.0.0
    # via requests
cycler==0.10.0
    # via matplotlib
future==0.18.2
    # via pysparkling
idna==2.10
    # via requests
kiwisolver==1.3.1
    # via matplotlib
matplotlib==3.3.3
    # via -r requirements.in
numpy==1.19.5
    # via
    #   matplotlib
    #   opencv-python
    #   openpifpaf
    #   scipy
    #   torch
    #   torchvision
opencv-python==4.3.0.36
    # via -r requirements.in
openpifpaf==0.11.9
    # via -r requirements.in
pillow==8.1.0
    # via
    #   matplotlib
    #   openpifpaf
    #   torchvision
pyparsing==2.4.7
    # via matplotlib
pysparkling==0.6.0
    # via openpifpaf
python-dateutil==2.8.1
    # via matplotlib
python-json-logger==2.0.1
    # via openpifpaf
requests==2.25.1
    # via pysparkling
scipy==1.6.0
    # via openpifpaf
six==1.15.0
    # via
    #   cycler
    #   python-dateutil
torch==1.7.1
    # via
    #   openpifpaf
    #   torchvision
torchvision==0.8.2
    # via openpifpaf
typing-extensions==3.7.4.3
    # via torch
urllib3==1.26.2
    # via requests
